# backend/optimosql/models.go
// Code generated by sqlc. DO NOT EDIT.
// versions:
//   sqlc v1.28.0

package optimosql

import (
	"github.com/jackc/pgx/v5/pgtype"
)

type Product struct {
	ID        int32
	Name      string
	Sku       pgtype.Text
	Brand     pgtype.Text
	CreatedAt pgtype.Timestamp
}

type PurchaseOrder struct {
	ID                    int32
	SupplierID            string
	OrderDate             pgtype.Date
	EstimatedDeliveryDate pgtype.Date
	CreatedAt             pgtype.Timestamp
	ProductID             int32
	Quantity              int32
}

type Sale struct {
	ID        int32
	ProductID int32
	Quantity  int32
	SaleDate  pgtype.Date
	SalePrice pgtype.Numeric
	CreatedAt pgtype.Timestamp
}