# openapi/apispec.yaml
openapi: 3.0.3
info:
  title: Optimo API
  version: 1.0.0
  description: API for managing sales and purchase orders for clothing brands.

servers:
  - url: http://localhost:8080

paths:
  /sales:
    get:
      summary: Get sales data aggregated by month
      description: Returns a list of sales aggregated by month. Optionally filter by product IDs.
      parameters:
        - in: query
          name: start_date
          schema:
            type: string
            format: date
          description: Filter sales starting from this date (inclusive)
        - in: query
          name: end_date
          schema:
            type: string
            format: date
          description: Filter sales up to this date (inclusive)         
        - in: query
          name: product_ids
          schema:
            type: array
            items:
              type: integer
          style: form
          explode: true
          description: Filter sales by one or more product IDs
      responses:
        '200':
          description: Successful response with aggregated sales
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/SalesData'

  /purchase-orders:
    get:
      summary: List all purchase orders
      description: Returns all purchase orders with supplier, product, and delivery info.
      responses:
        '200':
          description: A list of purchase orders
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/PurchaseOrder'

    post:
      summary: Create a new purchase order
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NewPurchaseOrder'
      responses:
        '201':
          description: Purchase order created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PurchaseOrder'

  /purchase-orders/{id}:
    delete:
      summary: Delete a purchase order
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        '204':
          description: Purchase order deleted

components:
  schemas:
    SalesData:
      type: object
      properties:
        year_month:
          type: string
        sales_turnover:
          type: number
          format: float
        inventory_turnover:
          type: integer
        product_id:
          type: integer

    PurchaseOrder:
      type: object
      properties:
        id:
          type: integer
        supplier_id:
          type: string
        order_date:
          type: string
          format: date
        estimated_delivery_date:
          type: string
          format: date
        product_id:
          type: integer
        quantity:
          type: integer
        created_at:
          type: string
          format: date-time

    NewPurchaseOrder:
      type: object
      required:
        - supplier_id
        - order_date
        - product_id
        - quantity
      properties:
        supplier_id:
          type: string
        order_date:
          type: string
          format: date
        estimated_delivery_date:
          type: string
          format: date
        product_id:
          type: integer
        quantity:
          type: integer